requirements.txt
channels==4.3.1
daphne==4.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-knox==5.0.2
djangorestframework==3.16.1
drf-spectacular==0.28.0
google-auth==2.40.3
Markdown==3.8.2
pillow==11.3.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4